Fix wrong generated libsvm data format
Due to the defacts of the Python script, it will only write the
label once and then write all the index:value pair without a new
line between two records, although they are under the same label.
svm-train will only tell me the first line where it encounters
error, which is awful as the file is so large that I can hardly
locate the bug's position. But the tool checkdata.py finally shot
the problem.
What's more, the appending writing method makes the train file
extremely large, which should be noticed next time!